+++ processor/src/execution/operations/fri_ops/mod.rs
@@ -57,7 +57,7 @@ where
     // the power of the domain generator which can be used to determine current domain value x
     let poe = processor.stack().get(10);
     if poe.is_zero() {
-        return Err(OperationError::InvalidFriDomainGenerator);
+        return Err(OperationError::FriError("domain size was 0".into()));
     }
     // the result of the previous layer folding
     let prev_value = {
@@ -76,12 +76,17 @@ where
 
     // --- make sure the previous folding was done correctly --------------
     if domain_segment > 3 {
-        return Err(OperationError::InvalidFriDomainSegment(domain_segment));
+        return Err(OperationError::FriError(format!(
+            "domain segment value cannot exceed 3, but was {domain_segment}"
+        )));
     }
 
     let d_seg = domain_segment as usize;
     if query_values[d_seg] != prev_value {
-        return Err(OperationError::InvalidFriLayerFolding(prev_value, query_values[d_seg]));
+        return Err(OperationError::FriError(format!(
+            "degree-respecting projection is inconsistent: expected {} but was {}",
+            prev_value, query_values[d_seg]
+        )));
     }
 
     // --- fold query values ----------------------------------------------